Refactor reader.go
