+++ internal/storage/metricstore/elasticsearch/reader.go
@@ -20,8 +20,6 @@ import (
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/metricstore"
 )
 
-var ErrNotImplemented = errors.New("metrics querying is currently not implemented yet")
-
 const minStep = time.Millisecond
 
 // MetricsReader orchestrates metrics queries by:
@@ -102,8 +100,7 @@ func (r MetricsReader) GetLatencies(ctx context.Context, params *metricstore.Lat
 	}
 
 	// Process the raw aggregation value to calculate latencies (ms)
-	const lookback = 1 // only current value
-	return applySlidingWindow(rawMetricFamily, lookback, scaleToMillisAndRound), nil
+	return ProcessLatencies(rawMetricFamily), nil
 }
 
 // GetCallRates retrieves call rate metrics
@@ -132,10 +129,7 @@ func (r MetricsReader) GetCallRates(ctx context.Context, params *metricstore.Cal
 		return nil, err
 	}
 
-	// Process and return results
-	processedMetricFamily := calcCallRate(rawMetricFamily, params.BaseQueryParameters)
-	// Trim results to original time range
-	return trimMetricPointsBefore(processedMetricFamily, timeRange.startTimeMillis), nil
+	return ProcessCallRates(rawMetricFamily, params.BaseQueryParameters, timeRange), nil
 }
 
 // GetErrorRates retrieves error rate metrics
@@ -159,190 +153,22 @@ func (r MetricsReader) GetErrorRates(ctx context.Context, params *metricstore.Er
 	}
 	// Convert search results into raw metric family using translator
 	translator := NewTranslator(bucketsToCallRate)
-	rawErrorsMetricFamily, err := translator.ToDomainMetricsFamily(metricsParams, searchResult)
+	rawErrorsMetrics, err := translator.ToDomainMetricsFamily(metricsParams, searchResult)
 	if err != nil {
 		return nil, err
 	}
-
-	// Process and return results (using the same calculate method as GetCallRates)
-	errorsMetrics := calcCallRate(rawErrorsMetricFamily, params.BaseQueryParameters)
 
 	callRateMetrics, err := r.GetCallRates(ctx, &metricstore.CallRateQueryParameters{BaseQueryParameters: params.BaseQueryParameters})
 	if err != nil {
 		return nil, err
 	}
 
-	// Calculate error rates as errors/sec divided by calls/sec
-	errorRateMetrics := calcErrorRates(trimMetricPointsBefore(errorsMetrics, timeRange.startTimeMillis), callRateMetrics)
-
-	// Trim results to original time range
-	return errorRateMetrics, nil
-}
-
-// calcErrorRates computes error rates by dividing error metrics by call metrics
-func calcErrorRates(errorMetrics, callMetrics *metrics.MetricFamily) *metrics.MetricFamily {
-	result := &metrics.MetricFamily{
-		Name:    errorMetrics.Name,
-		Type:    metrics.MetricType_GAUGE,
-		Help:    errorMetrics.Help,
-		Metrics: make([]*metrics.Metric, 0, len(errorMetrics.Metrics)),
-	}
-
-	for i, errorMetric := range errorMetrics.Metrics {
-		if i >= len(callMetrics.Metrics) {
-			break
-		}
-
-		callMetric := callMetrics.Metrics[i]
-		metricPoints := calculateErrorRatePoints(errorMetric.MetricPoints, callMetric.MetricPoints)
-
-		result.Metrics = append(result.Metrics, &metrics.Metric{
-			Labels:       errorMetric.Labels,
-			MetricPoints: metricPoints,
-		})
-	}
-
-	return result
-}
-
-// calculateErrorRatePoints computes error rates for corresponding metric points
-func calculateErrorRatePoints(errorPoints, callPoints []*metrics.MetricPoint) []*metrics.MetricPoint {
-	metricPoints := make([]*metrics.MetricPoint, 0, len(errorPoints))
-
-	for j, errorPoint := range errorPoints {
-		if j >= len(callPoints) {
-			break
-		}
-
-		callPoint := callPoints[j]
-		value := calculateErrorRateValue(errorPoint, callPoint)
-
-		metricPoints = append(metricPoints, &metrics.MetricPoint{
-			Timestamp: errorPoint.Timestamp,
-			Value:     toDomainMetricPointValue(value),
-		})
-	}
-
-	return metricPoints
-}
-
-// calculateErrorRateValue computes the error rate for a single point
-func calculateErrorRateValue(errorPoint, callPoint *metrics.MetricPoint) float64 {
-	errorValue := errorPoint.GetGaugeValue().GetDoubleValue()
-	callValue := callPoint.GetGaugeValue().GetDoubleValue()
-
-	if math.IsNaN(errorValue) && !math.IsNaN(callValue) {
-		return 0.0
-	}
-
-	if math.IsNaN(errorValue) || math.IsNaN(callValue) {
-		return math.NaN()
-	}
-	return errorValue / callValue
+	return ProcessErrorRates(rawErrorsMetrics, callRateMetrics, params.BaseQueryParameters, timeRange), nil
 }
 
 // GetMinStepDuration returns the minimum step duration.
 func (MetricsReader) GetMinStepDuration(_ context.Context, _ *metricstore.MinStepDurationQueryParameters) (time.Duration, error) {
 	return minStep, nil
-}
-
-// trimMetricPointsBefore removes metric points older than startMillis from each metric in the MetricFamily.
-func trimMetricPointsBefore(mf *metrics.MetricFamily, startMillis int64) *metrics.MetricFamily {
-	for _, metric := range mf.Metrics {
-		points := metric.MetricPoints
-		// Find first index where point >= startMillis
-		cutoff := 0
-		for ; cutoff < len(points); cutoff++ {
-			point := points[cutoff]
-			pointMillis := point.Timestamp.Seconds*1000 + int64(point.Timestamp.Nanos)/1000000
-			if pointMillis >= startMillis {
-				break
-			}
-		}
-		// Slice the array starting from cutoff index
-		metric.MetricPoints = points[cutoff:]
-	}
-	return mf
-}
-
-// applySlidingWindow applies a given processing function over a moving window of metric points.
-// This is the core generic function that contains the shared logic.
-func applySlidingWindow(mf *metrics.MetricFamily, lookback int, processor func(window []*metrics.MetricPoint) float64) *metrics.MetricFamily {
-	for _, metric := range mf.Metrics {
-		points := metric.MetricPoints
-		if len(points) == 0 {
-			continue
-		}
-
-		processedPoints := make([]*metrics.MetricPoint, 0, len(points))
-
-		for i, currentPoint := range points {
-			// Define the start of the moving window, ensuring it's not out of bounds.
-			start := i - lookback + 1
-			if start < 0 {
-				start = 0
-			}
-			window := points[start : i+1]
-
-			// Delegate the specific calculation to the provided processor function.
-			resultValue := processor(window)
-
-			processedPoints = append(processedPoints, &metrics.MetricPoint{
-				Timestamp: currentPoint.Timestamp,
-				Value:     toDomainMetricPointValue(resultValue),
-			})
-		}
-		metric.MetricPoints = processedPoints
-	}
-	return mf
-}
-
-// calcCallRate defines the rate calculation logic and pass in applySlidingWindow.
-func calcCallRate(mf *metrics.MetricFamily, params metricstore.BaseQueryParameters) *metrics.MetricFamily {
-	lookback := int(math.Ceil(float64(params.RatePer.Milliseconds()) / float64(params.Step.Milliseconds())))
-	// Ensure lookback >= 1
-	lookback = int(math.Max(float64(lookback), 1))
-
-	windowSizeSeconds := float64(lookback) * params.Step.Seconds()
-
-	// rateCalculator is a closure that captures 'lookback' and 'windowSizeSeconds'.
-	// It implements the specific logic for calculating the rate.
-	rateCalculator := func(window []*metrics.MetricPoint) float64 {
-		// If the window is not "full" (i.e., we don't have enough preceding points
-		// to calculate a rate over the full 'lookback' period), return NaN.
-		if len(window) < lookback {
-			return math.NaN()
-		}
-
-		firstValue := window[0].GetGaugeValue().GetDoubleValue()
-		// If the first value in the full window is NaN, treat it as 0 for the rate calculation.
-		// This implies that if data was missing at the start of the window, we assume no contribution from that missing period.
-		if math.IsNaN(firstValue) {
-			firstValue = 0
-		}
-		lastValue := window[len(window)-1].GetGaugeValue().GetDoubleValue()
-		// If the current point (the last value in the window) is NaN, the rate cannot be defined.
-		// Propagate NaN to indicate missing data for the result point.
-		if math.IsNaN(lastValue) {
-			return math.NaN()
-		}
-
-		rate := (lastValue - firstValue) / windowSizeSeconds
-		return math.Round(rate*100) / 100
-	}
-
-	return applySlidingWindow(mf, lookback, rateCalculator)
-}
-
-func scaleToMillisAndRound(window []*metrics.MetricPoint) float64 {
-	if len(window) == 0 {
-		return math.NaN()
-	}
-
-	v := window[len(window)-1].GetGaugeValue().GetDoubleValue()
-	// Scale down the value (e.g., from microseconds to milliseconds)
-	resultValue := v / 1000.0
-	return math.Round(resultValue*100) / 100 // Round to 2 decimal places
 }
 
 // bucketsToPoints is a helper function for getting points value from ES AGG bucket
